requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bioc==2.1
boto3==1.35.18
botocore==1.35.18
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.7
click==8.1.7
coloredlogs==15.0.1
conllu==4.5.3
contourpy==1.2.1
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.0.0
deepdiff==8.0.1
Deprecated==1.2.14
dill==0.3.8
docopt==0.6.2
durationpy==0.7
emoji==2.13.2
exceptiongroup==1.2.1
fastapi==0.115.0
filelock==3.15.4
filetype==1.2.0
flair==0.14.0
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gdown==5.2.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
intervaltree==3.1.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.5
kubernetes==31.0.0
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.22
langchain-huggingface==0.0.3
langchain-nvidia-ai-endpoints==0.2.2
langchain-text-splitters==0.2.1
langdetect==1.0.9
langgraph==0.1.19
langgraph-checkpoint==1.0.11
langsmith==0.1.125
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.1
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpld3==0.5.10
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.19.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orderly-set==5.2.2
orjson==3.10.7
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
posthog==3.6.6
pptree==3.1
protobuf==4.25.5
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.2
pypdf==5.0.0
PyPika==0.48.9
pyproject_hooks==1.1.0
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-oxmsg==0.0.1
pytorch_revgrad==0.2.0
pytz==2024.1
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.0
scipy==1.13.1
segtok==1.5.11
semver==3.0.2
sentence-transformers==3.1.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sparse==0.15.4
SQLAlchemy==2.0.35
sqlitedict==2.1.0
starlette==0.38.6
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.6.0.dev20240918
torchaudio==2.5.0.dev20240918
torchvision==0.20.0.dev20240918
tqdm==4.66.4
transformer-smaller-training-vocab==0.4.0
transformers==4.44.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.13
unstructured-client==0.25.9
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Wikipedia-API==0.7.1
wrapt==1.16.0
xxhash==3.5.0
yarl==1.11.1
zipp==3.20.2
